Adicionar novos endpoints para gerenciamento de tokens e agendamento de mensagens, além de atualizar dependências no pyproject.toml e ajustar configuração do Docker Compose.
