+++ compose.yaml
@@ -1,9 +1,9 @@
 services:
   api-bot:
-    image: joaoprdo/bot-whatsapp:latest
-    #build:
-    #  context: .
-    #  dockerfile: Dockerfile
+    #image: joaoprdo/bot-whatsapp:latest
+    build:
+      context: .
+      dockerfile: Dockerfile
     ports:
       - "8000:8000"
     env_file:
