+++ pyproject.toml
@@ -11,6 +11,10 @@ fastapi = "^0.115.6"
 uvicorn = "^0.34.0"
 requests = "^2.32.3"
 python-dotenv = "^1.0.1"
+pytz = "^2024.2"
+apscheduler = "^3.11.0"
+sqlalchemy = "^2.0.36"
+pydantic = "^2.10.4"
 
 
 [build-system]
